+++ app/Resources/FOSUserBundle/views/Security/login_content.html.twig
@@ -18,13 +18,6 @@
     </div>
 
     <div class="login_btn_area">
-        <button type="submit" class="btn form-control login_btn" id="_submit" name="_submit">Login</button>
-        <div class="login_social">
-            <h5>Login With</h5>
-            <ul>
-                <li><a href="#"><i class="fa fa-facebook"></i></a></li>
-                <li><a href="#"><i class="fa fa-google-plus"></i></a></li>
-            </ul>
-        </div>
+        <button type="submit" class="btn form-control login_btn" style="width: 100%;" id="_submit" name="_submit">Login</button>
     </div>
 </form>
