+++ app/Resources/views/base.html.twig
@@ -297,7 +297,7 @@
         <section class="welcome_area">
             <div class="container">
                 <div class="welcome_title">
-                    <h3>Welcome to <span>Vero</span>Date</h3>
+                    <h3>Welcome to <span>My</span>Soulmate</h3>
                     <img src="{{ asset('member_assets/img/w-title-b.png') }}" alt="">
                 </div>
                 <div class="row">
@@ -342,7 +342,7 @@
                         <div class="item">
                             <div class="col-md-7">
                                 <div class="download_app_icon">
-                                    <h3>Download <span>Vero</span><span>Date</span> app</h3>
+                                    <h3>Download <span>My</span><span>Soulmate</span> app</h3>
                                     <h5>Free Available in All Store PlayStore, AppStore & Microsoft Store</h5>
                                     <ul>
                                         <li><a href="#"><i class="fa fa-android"></i></a></li>
@@ -416,60 +416,12 @@
             <div class="footer_widgets_area">
                 <div class="container">
                     <div class="row">
-                        <div class="col-md-3">
+                        <div class="col-md-12">
                             <aside class="f_widget">
                                 <div class="vero_widget">
-                                    <h4><span>Vero</span>Date</h4>
+                                    <h4><span>My</span>Soulmate</h4>
                                     <p>There are many variations of passag-es of Lorem Ipsum available, but the majority have suffered alteration in some form.</p>
                                     <p>If you are going to use a passage of Lorem Ipsum, you need to be sure there isn't anything .</p>
-                                </div>
-                            </aside>
-                        </div>
-                        <div class="col-md-3">
-                            <aside class="f_widget">
-                                <div class="f_widget_title">
-                                    <h3>Information</h3>
-                                </div>
-                                <div class="information_widget">
-                                    <ul>
-                                        <li><a href="#">About us</a></li>
-                                        <li><a href="#">Contact us</a></li>
-                                        <li><a href="#">Membership</a></li>
-                                        <li><a href="#">Private Policy</a></li>
-                                        <li><a href="#">Forum Support</a></li>
-                                    </ul>
-                                </div>
-                            </aside>
-                        </div>
-                        <div class="col-md-3">
-                            <aside class="f_widget">
-                                <div class="f_widget_title">
-                                    <h3>Recent posts</h3>
-                                </div>
-                                <div class="recent_post_widget">
-                                    <ul>
-                                        <li><a href="#">Blog Standard Post <span>14 Sep, 2016</span></a></li>
-                                        <li><a href="#">Blog Image Post <span>12 Sep, 2016</span></a></li>
-                                        <li><a href="#">BlogVideo Post <span>08 Sep, 2016</span></a></li>
-                                        <li><a href="#">Blog Audio  Post <span>03 Sep, 2016</span></a></li>
-                                        <li><a href="#">Blog Standard Post <span>16 Aug, 2016</span></a></li>
-                                    </ul>
-                                </div>
-                            </aside>
-                        </div>
-                        <div class="col-md-3">
-                            <aside class="f_widget">
-                                <div class="f_widget_title">
-                                    <h3>Newsletter Signup</h3>
-                                </div>
-                                <div class="newsletter_widget">
-                                    <p>Get Alert Directly Into Your Inbox After Each Post.</p>
-                                    <div class="input-group">
-                                        <input type="text" class="form-control" placeholder="Enter your email">
-                                        <span class="input-group-btn">
-                                                    <button class="btn btn-default" type="button"><i class="fa fa-paper-plane" aria-hidden="true"></i></button>
-                                                </span>
-                                    </div>
                                 </div>
                             </aside>
                         </div>
