+++ app/Resources/views/member_base.html.twig
@@ -104,9 +104,7 @@
                                <a href="{{ path('recommandations') }}">Recommandation</a>
                            </li>
                            <li class="dropdown submenu">
-                               <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">
-                                   <img src="{{ asset('member_assets/img/notification.png') }}" height="25">
-                               </a>
+
                                <ul class="dropdown-menu">
                                    {% block notification %}
                                    {% endblock %}
@@ -172,84 +170,15 @@
             <div class="footer_widgets_area">
                 <div class="container">
                     <div class="row">
-                        <div class="col-md-3">
+                        <div class="col-md-12">
                             <aside class="f_widget">
                                 <div class="vero_widget">
-                                    <h4><span>Vero</span>Date</h4>
+                                    <h4><span>My</span>Soulmate</h4>
                                     <p>There are many variations of passag-es of Lorem Ipsum available, but the majority have suffered alteration in some form.</p>
                                     <p>If you are going to use a passage of Lorem Ipsum, you need to be sure there isn't anything .</p>
                                 </div>
                             </aside>
                         </div>
-                        <div class="col-md-3">
-                            <aside class="f_widget">
-                                <div class="f_widget_title">
-                                    <h3>Information</h3>
-                                </div>
-                                <div class="information_widget">
-                                    <ul>
-                                        <li><a href="#">About us</a></li>
-                                        <li><a href="#">Contact us</a></li>
-                                        <li><a href="#">Membership</a></li>
-                                        <li><a href="#">Private Policy</a></li>
-                                        <li><a href="#">Forum Support</a></li>
-                                    </ul>
-                                </div>
-                            </aside>
-                        </div>
-                        <div class="col-md-3">
-                            <aside class="f_widget">
-                                <div class="f_widget_title">
-                                    <h3>Recent posts</h3>
-                                </div>
-                                <div class="recent_post_widget">
-                                    <ul>
-                                        <li><a href="#">Blog Standard Post <span>14 Sep, 2016</span></a></li>
-                                        <li><a href="#">Blog Image Post <span>12 Sep, 2016</span></a></li>
-                                        <li><a href="#">BlogVideo Post <span>08 Sep, 2016</span></a></li>
-                                        <li><a href="#">Blog Audio  Post <span>03 Sep, 2016</span></a></li>
-                                        <li><a href="#">Blog Standard Post <span>16 Aug, 2016</span></a></li>
-                                    </ul>
-                                </div>
-                            </aside>
-                        </div>
-                        <div class="col-md-3">
-                            <aside class="f_widget">
-                                <div class="f_widget_title">
-                                    <h3>Newsletter Signup</h3>
-                                </div>
-                                <div class="newsletter_widget">
-                                    <p>Get Alert Directly Into Your Inbox After Each Post.</p>
-                                    <div class="input-group">
-                                        <input type="text" class="form-control" placeholder="Enter your email">
-                                        <span class="input-group-btn">
-                                            <button class="btn btn-default" type="button"><i class="fa fa-paper-plane" aria-hidden="true"></i></button>
-                                        </span>
-                                    </div>
-                                </div>
-                            </aside>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="copyright">
-                <div class="copyright_left">
-                    <div class="copyright_text">
-                        <h4>Copyright © 2017. <a href="#">VeroDate</a> . all rights reserved</h4>
-                    </div>
-                </div>
-                <div class="copyright_right">
-                    <div class="copyright_social">
-                        <ul>
-                            <li><a href="#"><i class="fa fa-facebook" aria-hidden="true"></i></a></li>
-                            <li><a href="#"><i class="fa fa-twitter" aria-hidden="true"></i></a></li>
-                            <li><a href="#"><i class="fa fa-pinterest" aria-hidden="true"></i></a></li>
-                            <li><a href="#"><i class="fa fa-linkedin" aria-hidden="true"></i></a></li>
-                            <li><a href="#"><i class="fa fa-google-plus" aria-hidden="true"></i></a></li>
-                            <li><a href="#"><i class="fa fa-google-plus" aria-hidden="true"></i></a></li>
-                            <li><a href="#"><i class="fa fa-vimeo" aria-hidden="true"></i></a></li>
-                            <li><a href="#"><i class="fa fa-youtube-play" aria-hidden="true"></i></a></li>
-                        </ul>
                     </div>
                 </div>
             </div>
