+++ app/config/security.yml
@@ -55,6 +55,6 @@ security:
         - { path: ^/admin/login$, role: IS_AUTHENTICATED_ANONYMOUSLY }
         - { path: ^/admin/login_check, role: IS_AUTHENTICATED_ANONYMOUSLY }
         - { path: ^/admin/logout, role: IS_AUTHENTICATED_ANONYMOUSLY }
-#        - { path: ^/admin, role: ROLE_ADMIN }
+        - { path: ^/admin, role: ROLE_ADMIN }
         - { path: ^/member, role: ROLE_USER }
-#        - { path: ^/business, role: ROLE_BUSINESS }
+        - { path: ^/business, role: ROLE_BUSINESS }
