+++ src/Admin/BusinessBundle/Resources/views/Advert/lister.html.twig
@@ -1,7 +1,7 @@
 {% extends 'admin_base.html.twig' %}
 
 
-{% block title %}AdminBusinessBundle:Advert:Lister{% endblock %}
+{% block title %}Adverts | List{% endblock %}
 {% block stylesheets %}
 {% endblock %}
 {% block scripts %}
@@ -20,28 +20,13 @@
 
             {% endfor %}
         });
-        {% for m in pubs %}
-
-        $('#dlt{{ m.id }}').click(function (){
-            $('#a{{ m.id }}').remove();
-            $.ajax({
-                url: "{{path('supprimer_admin')}}",
-                type: "POST",
-                cache: false,
-                data: {
-                    "id" : {{ m.id }}
-                }
-
-            });
-        });
-        {% endfor %}
 
 
         $('#delete').click(function (){
             var id = $("#selected_post").data('id');
             $('#a'+id).remove();
             $.ajax({
-                url: "{{path('delete_post_business')}}",
+                url: "{{path('supprimer_admin')}}",
                 type: "POST",
                 cache: false,
                 data: {
@@ -112,7 +97,7 @@
         <div class="modal-dialog" role="document">
             <div class="modal-content">
                 <div class="modal-header">
-                    <h5 class="modal-title">Modal title</h5>
+                    <h5 class="modal-title">Delete this Advert ?</h5>
                     <button type="button" class="close" data-dismiss="modal" aria-label="Close">
                         <span aria-hidden="true">&times;</span>
                     </button>
@@ -127,6 +112,8 @@
             </div>
         </div>
     </div>
+    <div id="selected_post" data-id=""></div>
+
 
     <h1>Publicités actives : </h1>
 
@@ -150,7 +137,7 @@
 
                         {% for m in pubs %}
                                 <tr id="a{{ m.id }}">
-                                    <td><span class="label label-info">{{ m.business.firstname }}</span></td>
+                                    <td><span class="label label-info">{{ m.business.username }}</span></td>
                                     <td><span class="label label-info"> {{ m.content }}</span></td>
                                     <td><span class="label label-info">{{ m.reason }}</span></td>
                                     <td> <img   src="{{ asset('member_assets/images/') }}{{ m.photoUrl}}"
