+++ src/Admin/BusinessBundle/Resources/views/Advert/listerpromos.html.twig
@@ -101,6 +101,7 @@
                             <a  >Delete </a></button></td>
 
                 </tr>
+
             {% endfor %}
         {% endfor %}
