+++ src/Admin/HomeBundle/Controller/DefaultController.php
@@ -12,6 +12,6 @@ class DefaultController extends Controller
      */
     public function indexAction()
     {
-        return $this->render('AdminHomeBundle:Default:index.html.twig');
+        return $this->redirectToRoute('admin_questions');
     }
 }
