+++ src/Business/AdvertBundle/Controller/BusinessAdvertController.php
@@ -328,72 +328,6 @@ class BusinessAdvertController extends Controller
         return $this->redirect('' . $payment->getApprovalLink());
     }
 
-    /**
-     * @Route("/Increment" , name="increment")
-     */
-
-    public function IncrementAction(Request $request)
-    {
-        $id = $request->request->get('id');
-        $advert = new Advert();
-        $repo = $this->getDoctrine()->getRepository(Advert::class);
-        $advert = $repo->IncrementClickDQL($id);
-        $var1 = $this->getDoctrine()->getRepository(Advert::class)->find($id);
-        if ($var1->getClicks() == 50) {
-
-            $message = (new Swift_Message())
-                ->setSubject('MySoulmate | Add approved !')
-                ->setFrom('mysoulmatepi@gmail.com')
-                ->setTo($var1->getBusiness()->getEmail())
-                ->setBody(
-                    "Bonjour Monsieur " . $var1->getBusiness()->getFirstName() . " , Votre Publicité : " . $var1->getContent() . " vient d'achever 
-                    le seuil de 50 clicks ! ",
-
-                    'text/html'
-                );
-            $this->get('mailer')->send($message);
-        } else if ($var1->getClicks() == 100) {
-
-            $message = (new Swift_Message())
-                ->setSubject('MySoulmate | Add approved !')
-                ->setFrom('mysoulmatepi@gmail.com')
-                ->setTo($var1->getBusiness()->getEmail())
-                ->setBody(
-                    "Bonjour Monsieur " . $var1->getBusiness()->getFirstName() . " , Votre Publicité : " . $var1->getContent() . " vient d'achever 
-                    le seuil de 100 clicks ! ",
-
-                    'text/html'
-                );
-            $this->get('mailer')->send($message);
-        } else if ($var1->getClicks() == 150) {
-
-            $message = (new Swift_Message())
-                ->setSubject('MySoulmate | Add approved !')
-                ->setFrom('mysoulmatepi@gmail.com')
-                ->setTo($var1->getBusiness()->getEmail())
-                ->setBody(
-                    "Bonjour Monsieur " . $var1->getBusiness()->getFirstName() . " , Votre Publicité : " . $var1->getContent() . " vient d'achever 
-                    le seuil de 150 clicks ! ",
-
-                    'text/html'
-                );
-            $this->get('mailer')->send($message);
-        } else if ($var1->getClicks() == 200) {
-
-            $message = (new Swift_Message())
-                ->setSubject('MySoulmate | Add approved !')
-                ->setFrom('mysoulmatepi@gmail.com')
-                ->setTo($var1->getBusiness()->getEmail())
-                ->setBody(
-                    "Bonjour Monsieur " . $var1->getBusiness()->getFirstName() . " , Votre Publicité : " . $var1->getContent() . " vient d'achever 
-                    le seuil de 250 clicks ! ",
-
-                    'text/html'
-                );
-            $this->get('mailer')->send($message);
-        }
-        return new Response("nice done kid");
-    }
 
     /**
      * @Route("/find" , name="find")
