+++ src/Business/AdvertBundle/Resources/views/BusinessAdvert/ajouter.html.twig
@@ -101,7 +101,26 @@ $(document).ready(function () {
 
     });
 
-        });
+    $('#valider').click(function f(f) {
+        var pub = new Date(
+            $('#basebundle_advert_endDate_year').val()
+            ,
+            $('#basebundle_advert_endDate_month').val()-1
+            ,
+            $('#basebundle_advert_endDate_day').val());
+        var d = new Date();
+        var s = (pub.getTime()-d.getTime());
+        var diff = Math.ceil(s / (1000 * 3600 * 24));
+
+        if (diff<0)
+        {   f.preventDefault();
+            alert('Please insert a valid date ');
+        }
+
+
+    })   ;
+
+});
 </script>
     {{ form_start(form,  { 'id': 'formtwig' }) }}<br>
     <table id="table">
@@ -109,13 +128,13 @@ $(document).ready(function () {
         <tr>
             <td> Promotion code:  </td>
             <td>
-                <div> <input type="text" id="promcode" placeholder="promo code" >
+                <div> <input type="text" id="promcode" placeholder="Promo code" >
                     <div class="alert alert-danger alert-dismissible" role="alert" id="no" hidden="true">
                         <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-                        <i class="fa fa-times-circle"></i> This code is wrong ,insert another one . </div>
+                        <i class="fa fa-times-circle"></i> This code is wrong or expired ,use another one . </div>
                     <div class="alert alert-success alert-dismissible" role="alert" id="yes" hidden="true">
                         <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-                        <i class="fa fa-check-circle"></i> This code gives you 30% discount
+                        <i class="fa fa-check-circle"></i> This code gives you 30% discount.
                     </div>
                 </div>
 
@@ -139,7 +158,7 @@ $(document).ready(function () {
 
 
         <tr>
-            <td>end date</td>
+            <td>End date</td>
             <td>{{ form_widget(form.endDate)}}</td>
         </tr>
 
@@ -153,7 +172,7 @@ $(document).ready(function () {
             <td>{{ form_widget(form.position ) }}</td>
         </tr>
         <tr>
-            <td> durée</td> <td> <input type="text" id="durée" placeholder="0 Jours" disabled="true" ></td>
+            <td> durée</td> <td> <input type="text" id="durée" placeholder="0 Days" disabled="true" ></td>
         </tr>
         <tr>
             <td>prix :</td>
@@ -168,7 +187,7 @@ $(document).ready(function () {
         <tr>
 
             <td class="demo-button">  </td>
-            <td class="demo-button"> <input type="button" value="Calculer prix" id="calcul-prix"> {{ form_widget(form.Valider ,  { 'id': 'valider' })  }}</td>
+            <td class="demo-button"> <input type="button" value="Show price" id="calcul-prix"> {{ form_widget(form.Valider ,  { 'id': 'valider' })  }}</td>
         </tr>
