+++ src/Business/AdvertBundle/Resources/views/BusinessAdvert/bigpub.html.twig
@@ -1,7 +1,7 @@
 {% for m in bigs %}
    {% set menuUri %}https://{{ m.business.link }}{% endset %}
 
-   <section role="button" class="banner_area" style="background: url({% block banner_img %} {{ asset('member_assets/images/') }}{{ m.photoUrl}}{% endblock %}) no-repeat scroll center center; background-size: cover;" id="x">
+   <section  class="banner_area" style="background: url({% block banner_img %} {{ asset('member_assets/images/') }}{{ m.photoUrl}}{% endblock %}) no-repeat scroll center center; background-size: cover;" id="x">
        <div class="container">
            <div class="banner_content">
                <br><br><br><br><br><br><br><br><br><br><br><br><br>
