+++ src/Business/AdvertBundle/Resources/views/BusinessAdvert/sidepub.html.twig
@@ -6,11 +6,11 @@
             {% for m in sides %}
                 {% for i in [1..10]  %}
                     {% if m.videoUrl == null  %}
-                        <div class="row">
+                        <div class="row" >
                             <div class="blog_grid_inner">
                                 <div class="col-md-3 col-sm-6">
-                                    <div class="blog_grid_item">
-                                        <div class="blog_grid_img">
+                                    <div class="blog_grid_item" style="width: 280px">
+                                        <div class="blog_grid_img" >
                                             <img src="{{ asset('member_assets/') }}images/{{ m.photoUrl }}" alt="" style="width: 280px ; height : 250px">
                                             <div class="blog_share_area">
                                                 <a></a>
@@ -32,7 +32,7 @@
                         <div class="row">
                             <div class="blog_grid_inner">
                                     <div class="blog_grid_item">
-                                        <div class="blog_grid_img">
+                                        <div class="blog_grid_img" style="width: 280px">
                                             <img src="{{ asset('member_assets/') }}images/{{ m.photoUrl }}" alt="" style="width: 280px ; height : 250px">
                                             <div class="blog_share_area">
                                                 <a   id="{{ m.id }}-ad"  ><i class="fa fa-heart" aria-hidden="true" role="button"></i>{{ m.clicks }}</a>
